back/pyproject.toml
[project]
name = "topologicook"
version = "0.1.0"
description = "Cooking helper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "orjson>=3.10.7",
    "loguru>=0.7.2",
    "extruct>=0.17.0",
    "dacite>=1.8.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
